feat: update endpoints to additional check

+++ src/routes/profiles/index.ts
@@ -2,6 +2,7 @@ import { FastifyPluginAsyncJsonSchemaToTs } from '@fastify/type-provider-json-sc
 import { idParamSchema } from '../../utils/reusedSchemas';
 import { createProfileBodySchema, changeProfileBodySchema } from './schema';
 import type { ProfileEntity } from '../../utils/DB/entities/DBProfiles';
+import { validate } from "uuid";
 
 const plugin: FastifyPluginAsyncJsonSchemaToTs = async (
   fastify
@@ -20,7 +21,11 @@ const plugin: FastifyPluginAsyncJsonSchemaToTs = async (
       },
     },
     async function (request, reply): Promise<ProfileEntity | null> {
-      return await fastify.db.profiles.findOne({ key: 'id', equals: request.params.id });
+      const profile = await fastify.db.profiles.findOne({ key: 'id', equals: request.params.id });
+      if (profile === null) {
+        throw fastify.httpErrors.notFound();
+      }
+      return profile;
     }
   );
 
@@ -32,7 +37,23 @@ const plugin: FastifyPluginAsyncJsonSchemaToTs = async (
       },
     },
     async function (request, reply): Promise<ProfileEntity> {
-      return await fastify.db.profiles.create(request.body);
+      const user = await fastify.db.users.findOne({ key: 'id', equals: request.body.userId });
+      if (user === null) {
+        throw fastify.httpErrors.badRequest();
+      }
+
+      const memberType = await fastify.db.memberTypes.findOne({ key: 'id', equals: request.body.memberTypeId });
+      if (memberType === null) {
+        throw fastify.httpErrors.badRequest();
+      }
+
+      const profileByUserId = await fastify.db.profiles.findOne({ key: 'userId', equals: request.body.userId });
+      if (profileByUserId !== null) {
+        throw fastify.httpErrors.badRequest();
+      }
+
+      const profile = await fastify.db.profiles.create(request.body);
+      return profile;
     }
   );
 
@@ -44,7 +65,16 @@ const plugin: FastifyPluginAsyncJsonSchemaToTs = async (
       },
     },
     async function (request, reply): Promise<ProfileEntity> {
-      return await fastify.db.profiles.delete(request.params.id);
+      if (!validate(request.params.id)) {
+        throw fastify.httpErrors.badRequest();
+      }
+
+      const profile = await fastify.db.profiles.findOne({ key: 'id', equals: request.params.id });
+      if (profile === null) {
+        throw fastify.httpErrors.notFound();
+      }
+      await fastify.db.profiles.delete(request.params.id);
+      return profile;
     }
   );
 
@@ -57,7 +87,17 @@ const plugin: FastifyPluginAsyncJsonSchemaToTs = async (
       },
     },
     async function (request, reply): Promise<ProfileEntity> {
-      return await fastify.db.profiles.change(request.params.id, request.body);
+      if (!validate(request.params.id)) {
+        throw fastify.httpErrors.badRequest();
+      }
+      
+      const profile = await fastify.db.profiles.findOne({ key: 'id', equals: request.params.id })
+      if (profile === null) {
+        throw fastify.httpErrors.notFound();
+      }
+
+      const result = await fastify.db.profiles.change(request.params.id, request.body);
+      return result!;
     }
   );
 };
